Update positivenegative.py
positive?negative

+++ positivenegative.py
@@ -1,9 +1,9 @@
-#python code goes gere
+#python code goes here
 #python version :3.2.5
-num = float(input("enter the number:"))
+num = int(input("enter the number:"))
 if num > 0:
   print("positive number")
-elif num ==0:
+elif num == 0:
   print("zero")
- else:
+else:
   print("the number is negative")
